+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/methods/EthJsonRpcMethods.java
@@ -46,6 +46,7 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetStorageV
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetTransactionByBlockHashAndIndex;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetTransactionByBlockNumberAndIndex;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetTransactionByHash;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetTransactionBySenderAndNonce;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetTransactionCount;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetTransactionReceipt;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.EthGetUncleByBlockHashAndIndex;
@@ -160,6 +161,7 @@ public class EthJsonRpcMethods extends ApiGroupJsonRpcMethods {
             new EthGetTransactionByHash(blockchainQueries, transactionPool),
             new EthGetTransactionByBlockHashAndIndex(blockchainQueries),
             new EthGetTransactionByBlockNumberAndIndex(blockchainQueries),
+            new EthGetTransactionBySenderAndNonce(blockchainQueries),
             new EthGetTransactionCount(blockchainQueries, transactionPool),
             new EthGetTransactionReceipt(blockchainQueries, protocolSchedule),
             new EthUninstallFilter(filterManager),